Something that does type dispatching is working but hacky at the moment
Some refactoring is certainly in order to clean up some of the different
ideas that are currently floating around in the code at the moment.  The model
now includes some more advanced object tree/graph dynamic dependencies that
need to be thought out a bit more.  DependentField's implementation turned
out particularly hacky.

In addition, a bug was discovered (there is a failing unit test with this
release) that shows that state is being shared between message instances.
This needs to be fixed (maybe by using prototypes for efficient object
instantatiation with the required class-level state).