refactor: allow for just-in-time length checks via bytes_required on fields
This allows for better protocol handling (still being developed) in that we
we wait to collect all the required data for a field before trying to parse
the field (leading to awkward/slow error catching).  This also allows us to
better control IO as we are just doing read()'s (when parsing) in a single
location.