Add support for greedy Payload fields
This change adds support for greedy payloads.  There can be one per message
and boxed or basic (just a payload at the end of the packet).  A pivot
strategy around the greedy field is used in order to support the boxed case.
Greedy payloads only work in cases where the context in which the message is
going to be used has packet boundaries that are already known.  This turns
out to be the case for many messages out there when dealing with multiple
layers in a protocol.