rename: pacman is now suitcase
After some deliberation, it was decided that pacman was too likely
to cause confusion as it is the name of the package manager under
Arch linux.  The name suitcase is related to packing/unpacking
in some fashion, has some character, isn't too long, and should
be pretty easy to google for.  I was unable to find any other major
python packages with this name and the name is available on pypi.