fields: 100% code coverage and bug fix for CRC calculation
This change adds tests in order to get to 100% statement coverage
with the unit tests.  In the course of testing, a bug was discovered
with the logic for a CRC field with a negative index specified for
the end value.  That bug is fixed and now all of the tests pass
and coverage shows that all statements in the library are covered by
the tests.