Handle receiving a magic start sequence in two separate parts
Previously if the StreamProtocolHandler was fed bytes that contained only
part of the Magic field that specifies the start of a message it would
throw away that start as it couldn't find the whole field.  Now, it is
aware of that possibility and keeps enough bytes around that it won't miss
a Magic field.

Signed-off-by: Zach Varberg <Zachary.Varberg@Digi.com>