fields: use polymorphism for effective substructure determination
This code is functionally equivalent to the previous version for
ConditionalField, but we now properly consider other fields that
could wrap SubstructureField to be SubstructureField to be such
(e.g. LengthField, TypeField).

This change also adds **kwargs to unpack() on all fields for consistency
and expansion as needed in the future.

Signed-off-by: Paul Osborne <Paul.Osborne@digi.com>