Fix unpack of Substructure or FieldArray when using a DependentField
At least with a LengthField, the referred field does not seem to automatically
update when used in the substructure. But at least for unpacking,
this fix does seem to work.

Basically the cause of issues before was that SubstructureField
and FieldArray did not set _parent on the substructure,
whereas DispatchTarget and ConditionalField did.